added a contact person to every project, added some more project tags, modified projects to also show month of duration, modified projects.markdown to sort projects to display not only by year but also by month, added filechecks to generate_publ_html plugin so execution takes less time, extracted publications from person 2's publication list to display on person 1's publication list

+++ _plugins/generate_publ_html.rb
@@ -1,4 +1,4 @@
-=begin
+puts "Running plugin generate_publ_html..."
 # Specifies the hook. This determines in which step of the build process this script will be run.
 Jekyll::Hooks.register :people, :pre_render do |person|
     name = person.data['lastname']
@@ -12,7 +12,10 @@ Jekyll::Hooks.register :people, :pre_render do |person|
             'ue'
         end
     end
-    File.open("_includes/publications/" + newname.downcase + ".html", 'w') {|f| f.write("<p>No publications available.</p>")}
+    if !File.exist?("_includes/publications/" + newname.downcase + ".html")
+        puts "Creating file" + newname.downcase + ".html"
+        File.open("_includes/publications/" + newname.downcase + ".html", 'w') {|f| f.write("<p>No publications available.</p>")}
+    end
 end
 
 Jekyll::Hooks.register :projects, :pre_render do |project|
@@ -27,6 +30,8 @@ Jekyll::Hooks.register :projects, :pre_render do |project|
             'ue'
         end
     end
-    File.open("_includes/publications/" + p_newabbr + ".html", 'w') {|f| f.write("<p>No publications available.</p>")}
+    if !File.exist?("_includes/publications/" + p_newabbr + ".html")
+        puts "Creating file" + p_newabbr + ".html"
+        File.open("_includes/publications/" + p_newabbr + ".html", 'w') {|f| f.write("<p>No publications available.</p>")}
+    end
 end
-=end
